fix page loaders
